+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - TreeView `lazyPopulate` alternative to `populate` that lazily creates children as needed
 - TreeItem nesting extension functions (https://github.com/edvin/tornadofx/issues/134)
 - TableView `selectWhere()`, `moveToTopWhere()` and `BottomWhere()` (https://github.com/edvin/tornadofx/issues/134)
+- Group builder `group`
 
 ### Changed